fixing build script

+++ deploy-ci.sh
@@ -1,40 +1,29 @@
 #!/bin/sh
 
-setup_git() {
-  git config --global user.email "kbl.bumigora@gmail.com"
-  git config --global user.name "KBL Bumigora"
-}
+if [[ $(git status -s) ]]
+then
+    echo "The working directory is dirty. Please commit any pending changes."
+    exit 1;
+fi
 
-build_website() {
-  echo "Deleting old publication"
-  rm -rf public
-  mkdir public
-  git worktree prune
-  rm -rf .git/worktrees/public/
+echo "Deleting old publication"
+rm -rf public
+mkdir public
+git worktree prune
+rm -rf .git/worktrees/public/
 
-  echo "Checking out gh-pages branch into public"
-  git worktree add -B gh-pages public origin/gh-pages
+echo "Checking out gh-pages branch into public"
+git worktree add -B gh-pages public origin/gh-pages
 
-  echo "Removing existing files"
-  rm -rf public/*
+echo "Removing existing files"
+rm -rf public/*
 
-  echo "Generating site..."
-  hugo
-  cd public
-}
+echo "Generating site"
+hugo
 
-commit_website_files() {
-  #git checkout -b gh-pages
-  git add --all
-  git commit --message "Travis build: $TRAVIS_BUILD_NUMBER"
-}
-
-upload_files() {
-  git remote add origin-pages https://${GITHUB_TOKEN}@github.com/kblbumigora/website.git > /dev/null 2>&1
-  git push --force --set-upstream origin-pages gh-pages 
-}
-
-setup_git
-build_website
-commit_website_files
-upload_files
+echo "Updating gh-pages branch"
+cd public
+git add --all
+git commit -m "Publishing to gh-pages (deploy.sh)"
+git remote add origin-pages https://${GITHUB_TOKEN}@github.com/kblbumigora/website.git > /dev/null 2>&1
+git push --force --set-upstream origin-pages gh-pages
